zl1: Setup the second-step input boost
This reworks the input boost configuration for
second-step input boost.

Step 1 represents a strong frequency boost
to prepare for the expected jump in CPU load.

After step 1 is completed, step 2 gets activated.

Step 2 represents a weak frequency boost. It lasts
for a longer time than step 1.

We use an asynchronous frequency configuration
scheme for the little and big cluster to account
for the natural frequency scaling in real-world
cases based on statistics.

This is the config that gets applied:

Step 1:
   Duration: 90ms
   Frequency (Little): 1286.4Mhz
   Frequency (Big): 1363.2Mhz

Step 2:
   Duration: 150ms
   Frequency (Little): 1132.8Mhz
   Frequency (Big): 1209.6Mhz

The complete input boost will take a minimum
of 240ms in the perfect case and a maximum of
270ms in the worst case (workqueue delays).

Change-Id: Ib6f0f00b0a67cba72c01681688113c2aed6b145c
Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>