zl1: properly declare more ACDB IDs for headset usecases
Switching to stereo speaker phone calls to enhance the audio experience
during calls.

Some ACDBs needed to be changed for this to work and also new ones
needed to be added to account for the usage of the Type-C adapter.

This also improves some parameters to support the infamous Type-C adapter.

Change-Id: Id205aeb81e67cc3c0bb2a20b62ff6b0a72e144ea
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>