zl1: fine-tuning audio configuration for better quality
Finally this should do it, the ULL problem was fixed by the blobs
update, the ACDB mapping had nothing to do with it.

Since we're getting back to EUI 23S audio mapping, given the previous
Zeusis changes, we'd rather keep most of them because of the current
LA.UM Kernel for compatibility.

After all, this change should enhance the audio quality by setting the
right ACDB IDs. We're also leveling up the audio via codecs and amplifier.

Change-Id: Iacd2c5f24b6495014cb07461ec860d5bc4cabe61
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>